requirements.txt
requests==2.21.0
gevent==1.3.7
demjson==2.2.4
numpy==1.17.3
pyquery==1.4.0
opencv_python==4.1.0.25
chardet==3.0.4
selenium==3.141.0
Flask==1.0.2
retrying==1.3.3
pynput==1.4.4
redis==3.2.1
PyExecJS==1.5.1
pycrypto==2.6.1
Pillow==6.2.1
beautifulsoup4==4.8.1
pyppeteer==0.0.25
